convert Worker env hash type from size_t to int

+++ src/ray/raylet/worker.cc
@@ -118,7 +118,7 @@ const std::unordered_set<TaskID> &Worker::GetBlockedTaskIds() const {
 
 const JobID &Worker::GetAssignedJobId() const { return assigned_job_id_; }
 
-size_t Worker::GetRuntimeEnvHash() const { return runtime_env_hash_; }
+int Worker::GetRuntimeEnvHash() const { return runtime_env_hash_; }
 
 void Worker::AssignActorId(const ActorID &actor_id) {
   RAY_CHECK(actor_id_.IsNil())
